backend/requirements.txt
fastapi
sqlalchemy
pymysql
fpdf2
uvicorn